Move row manipulation logic into RowManager class.
